My compilator issues warning. Uwe, please check that this was intended.
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@35491 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/frontends/qt4/GuiPainter.cpp
@@ -83,7 +83,7 @@ void GuiPainter::setQPainterPen(QColor const & col,
 		case line_onoffdash: pen.setStyle(Qt::DotLine); break;
 	}
 
-	pen.setWidth(lw);
+	pen.setWidth((int)lw);
 
 	setPen(pen);
 }
